+++ compiler-plugin/src/main/java/scalapb/options/compiler/Scalapb.java
@@ -332,6 +332,29 @@ public final class Scalapb {
 
     /**
      * <pre>
+     * If true, then source-code info information will be included in the
+     * generated code - normally the source code info is cleared out to reduce
+     * code size.  The source code info is useful for extracting source code
+     * location from the descriptors as well as comments.
+     * </pre>
+     *
+     * <code>optional bool retain_source_code_info = 13;</code>
+     */
+    boolean hasRetainSourceCodeInfo();
+    /**
+     * <pre>
+     * If true, then source-code info information will be included in the
+     * generated code - normally the source code info is cleared out to reduce
+     * code size.  The source code info is useful for extracting source code
+     * location from the descriptors as well as comments.
+     * </pre>
+     *
+     * <code>optional bool retain_source_code_info = 13;</code>
+     */
+    boolean getRetainSourceCodeInfo();
+
+    /**
+     * <pre>
      * For use in tests only. Inhibit Java conversions even when when generator parameters
      * request for it.
      * </pre>
@@ -374,6 +397,7 @@ public final class Scalapb {
       objectName_ = "";
       scope_ = 0;
       lenses_ = true;
+      retainSourceCodeInfo_ = false;
       testOnlyNoJavaConversions_ = false;
     }
 
@@ -479,8 +503,13 @@ public final class Scalapb {
               lenses_ = input.readBool();
               break;
             }
-            case 800008: {
+            case 104: {
               bitField0_ |= 0x00000400;
+              retainSourceCodeInfo_ = input.readBool();
+              break;
+            }
+            case 800008: {
+              bitField0_ |= 0x00000800;
               testOnlyNoJavaConversions_ = input.readBool();
               break;
             }
@@ -1087,6 +1116,35 @@ public final class Scalapb {
       return lenses_;
     }
 
+    public static final int RETAIN_SOURCE_CODE_INFO_FIELD_NUMBER = 13;
+    private boolean retainSourceCodeInfo_;
+    /**
+     * <pre>
+     * If true, then source-code info information will be included in the
+     * generated code - normally the source code info is cleared out to reduce
+     * code size.  The source code info is useful for extracting source code
+     * location from the descriptors as well as comments.
+     * </pre>
+     *
+     * <code>optional bool retain_source_code_info = 13;</code>
+     */
+    public boolean hasRetainSourceCodeInfo() {
+      return ((bitField0_ & 0x00000400) == 0x00000400);
+    }
+    /**
+     * <pre>
+     * If true, then source-code info information will be included in the
+     * generated code - normally the source code info is cleared out to reduce
+     * code size.  The source code info is useful for extracting source code
+     * location from the descriptors as well as comments.
+     * </pre>
+     *
+     * <code>optional bool retain_source_code_info = 13;</code>
+     */
+    public boolean getRetainSourceCodeInfo() {
+      return retainSourceCodeInfo_;
+    }
+
     public static final int TEST_ONLY_NO_JAVA_CONVERSIONS_FIELD_NUMBER = 100001;
     private boolean testOnlyNoJavaConversions_;
     /**
@@ -1098,7 +1156,7 @@ public final class Scalapb {
      * <code>optional bool test_only_no_java_conversions = 100001;</code>
      */
     public boolean hasTestOnlyNoJavaConversions() {
-      return ((bitField0_ & 0x00000400) == 0x00000400);
+      return ((bitField0_ & 0x00000800) == 0x00000800);
     }
     /**
      * <pre>
@@ -1163,6 +1221,9 @@ public final class Scalapb {
         output.writeBool(12, lenses_);
       }
       if (((bitField0_ & 0x00000400) == 0x00000400)) {
+        output.writeBool(13, retainSourceCodeInfo_);
+      }
+      if (((bitField0_ & 0x00000800) == 0x00000800)) {
         output.writeBool(100001, testOnlyNoJavaConversions_);
       }
       unknownFields.writeTo(output);
@@ -1228,6 +1289,10 @@ public final class Scalapb {
           .computeBoolSize(12, lenses_);
       }
       if (((bitField0_ & 0x00000400) == 0x00000400)) {
+        size += com.google.protobuf.CodedOutputStream
+          .computeBoolSize(13, retainSourceCodeInfo_);
+      }
+      if (((bitField0_ & 0x00000800) == 0x00000800)) {
         size += com.google.protobuf.CodedOutputStream
           .computeBoolSize(100001, testOnlyNoJavaConversions_);
       }
@@ -1300,6 +1365,11 @@ public final class Scalapb {
         result = result && (getLenses()
             == other.getLenses());
       }
+      result = result && (hasRetainSourceCodeInfo() == other.hasRetainSourceCodeInfo());
+      if (hasRetainSourceCodeInfo()) {
+        result = result && (getRetainSourceCodeInfo()
+            == other.getRetainSourceCodeInfo());
+      }
       result = result && (hasTestOnlyNoJavaConversions() == other.hasTestOnlyNoJavaConversions());
       if (hasTestOnlyNoJavaConversions()) {
         result = result && (getTestOnlyNoJavaConversions()
@@ -1369,6 +1439,11 @@ public final class Scalapb {
         hash = (37 * hash) + LENSES_FIELD_NUMBER;
         hash = (53 * hash) + com.google.protobuf.Internal.hashBoolean(
             getLenses());
+      }
+      if (hasRetainSourceCodeInfo()) {
+        hash = (37 * hash) + RETAIN_SOURCE_CODE_INFO_FIELD_NUMBER;
+        hash = (53 * hash) + com.google.protobuf.Internal.hashBoolean(
+            getRetainSourceCodeInfo());
       }
       if (hasTestOnlyNoJavaConversions()) {
         hash = (37 * hash) + TEST_ONLY_NO_JAVA_CONVERSIONS_FIELD_NUMBER;
@@ -1532,8 +1607,10 @@ public final class Scalapb {
         bitField0_ = (bitField0_ & ~0x00000400);
         lenses_ = true;
         bitField0_ = (bitField0_ & ~0x00000800);
-        testOnlyNoJavaConversions_ = false;
+        retainSourceCodeInfo_ = false;
         bitField0_ = (bitField0_ & ~0x00001000);
+        testOnlyNoJavaConversions_ = false;
+        bitField0_ = (bitField0_ & ~0x00002000);
         return this;
       }
 
@@ -1614,6 +1691,10 @@ public final class Scalapb {
         result.lenses_ = lenses_;
         if (((from_bitField0_ & 0x00001000) == 0x00001000)) {
           to_bitField0_ |= 0x00000400;
+        }
+        result.retainSourceCodeInfo_ = retainSourceCodeInfo_;
+        if (((from_bitField0_ & 0x00002000) == 0x00002000)) {
+          to_bitField0_ |= 0x00000800;
         }
         result.testOnlyNoJavaConversions_ = testOnlyNoJavaConversions_;
         result.bitField0_ = to_bitField0_;
@@ -1720,6 +1801,9 @@ public final class Scalapb {
         }
         if (other.hasLenses()) {
           setLenses(other.getLenses());
+        }
+        if (other.hasRetainSourceCodeInfo()) {
+          setRetainSourceCodeInfo(other.getRetainSourceCodeInfo());
         }
         if (other.hasTestOnlyNoJavaConversions()) {
           setTestOnlyNoJavaConversions(other.getTestOnlyNoJavaConversions());
@@ -2733,6 +2817,66 @@ public final class Scalapb {
         return this;
       }
 
+      private boolean retainSourceCodeInfo_ ;
+      /**
+       * <pre>
+       * If true, then source-code info information will be included in the
+       * generated code - normally the source code info is cleared out to reduce
+       * code size.  The source code info is useful for extracting source code
+       * location from the descriptors as well as comments.
+       * </pre>
+       *
+       * <code>optional bool retain_source_code_info = 13;</code>
+       */
+      public boolean hasRetainSourceCodeInfo() {
+        return ((bitField0_ & 0x00001000) == 0x00001000);
+      }
+      /**
+       * <pre>
+       * If true, then source-code info information will be included in the
+       * generated code - normally the source code info is cleared out to reduce
+       * code size.  The source code info is useful for extracting source code
+       * location from the descriptors as well as comments.
+       * </pre>
+       *
+       * <code>optional bool retain_source_code_info = 13;</code>
+       */
+      public boolean getRetainSourceCodeInfo() {
+        return retainSourceCodeInfo_;
+      }
+      /**
+       * <pre>
+       * If true, then source-code info information will be included in the
+       * generated code - normally the source code info is cleared out to reduce
+       * code size.  The source code info is useful for extracting source code
+       * location from the descriptors as well as comments.
+       * </pre>
+       *
+       * <code>optional bool retain_source_code_info = 13;</code>
+       */
+      public Builder setRetainSourceCodeInfo(boolean value) {
+        bitField0_ |= 0x00001000;
+        retainSourceCodeInfo_ = value;
+        onChanged();
+        return this;
+      }
+      /**
+       * <pre>
+       * If true, then source-code info information will be included in the
+       * generated code - normally the source code info is cleared out to reduce
+       * code size.  The source code info is useful for extracting source code
+       * location from the descriptors as well as comments.
+       * </pre>
+       *
+       * <code>optional bool retain_source_code_info = 13;</code>
+       */
+      public Builder clearRetainSourceCodeInfo() {
+        bitField0_ = (bitField0_ & ~0x00001000);
+        retainSourceCodeInfo_ = false;
+        onChanged();
+        return this;
+      }
+
       private boolean testOnlyNoJavaConversions_ ;
       /**
        * <pre>
@@ -2743,7 +2887,7 @@ public final class Scalapb {
        * <code>optional bool test_only_no_java_conversions = 100001;</code>
        */
       public boolean hasTestOnlyNoJavaConversions() {
-        return ((bitField0_ & 0x00001000) == 0x00001000);
+        return ((bitField0_ & 0x00002000) == 0x00002000);
       }
       /**
        * <pre>
@@ -2765,7 +2909,7 @@ public final class Scalapb {
        * <code>optional bool test_only_no_java_conversions = 100001;</code>
        */
       public Builder setTestOnlyNoJavaConversions(boolean value) {
-        bitField0_ |= 0x00001000;
+        bitField0_ |= 0x00002000;
         testOnlyNoJavaConversions_ = value;
         onChanged();
         return this;
@@ -2779,7 +2923,7 @@ public final class Scalapb {
        * <code>optional bool test_only_no_java_conversions = 100001;</code>
        */
       public Builder clearTestOnlyNoJavaConversions() {
-        bitField0_ = (bitField0_ & ~0x00001000);
+        bitField0_ = (bitField0_ & ~0x00002000);
         testOnlyNoJavaConversions_ = false;
         onChanged();
         return this;
@@ -8889,7 +9033,7 @@ public final class Scalapb {
   static {
     java.lang.String[] descriptorData = {
       "\n\rscalapb.proto\022\007scalapb\032 google/protobu" +
-      "f/descriptor.proto\"\230\003\n\016ScalaPbOptions\022\024\n" +
+      "f/descriptor.proto\"\271\003\n\016ScalaPbOptions\022\024\n" +
       "\014package_name\030\001 \001(\t\022\024\n\014flat_package\030\002 \001(" +
       "\010\022\016\n\006import\030\003 \003(\t\022\020\n\010preamble\030\004 \003(\t\022\023\n\013s" +
       "ingle_file\030\005 \001(\010\022\035\n\025no_primitive_wrapper" +
@@ -8897,32 +9041,33 @@ public final class Scalapb {
       "ollection_type\030\010 \001(\t\022\037\n\027preserve_unknown" +
       "_fields\030\t \001(\010\022\023\n\013object_name\030\n \001(\t\0223\n\005sc" +
       "ope\030\013 \001(\0162$.scalapb.ScalaPbOptions.Optio" +
-      "nsScope\022\024\n\006lenses\030\014 \001(\010:\004true\022\'\n\035test_on" +
-      "ly_no_java_conversions\030\241\215\006 \001(\010\"%\n\014Option" +
-      "sScope\022\010\n\004FILE\020\000\022\013\n\007PACKAGE\020\001\"~\n\016Message" +
-      "Options\022\017\n\007extends\030\001 \003(\t\022\031\n\021companion_ex" +
-      "tends\030\002 \003(\t\022\023\n\013annotations\030\003 \003(\t\022\014\n\004type" +
-      "\030\004 \001(\t\022\035\n\025companion_annotations\030\005 \003(\t\"\224\001" +
-      "\n\014FieldOptions\022\014\n\004type\030\001 \001(\t\022\022\n\nscala_na" +
-      "me\030\002 \001(\t\022\027\n\017collection_type\030\003 \001(\t\022\020\n\010key" +
-      "_type\030\004 \001(\t\022\022\n\nvalue_type\030\005 \001(\t\022\023\n\013annot" +
-      "ations\030\006 \003(\t\022\016\n\006no_box\030\036 \001(\010\"G\n\013EnumOpti" +
-      "ons\022\017\n\007extends\030\001 \003(\t\022\031\n\021companion_extend" +
-      "s\030\002 \003(\t\022\014\n\004type\030\003 \001(\t\"#\n\020EnumValueOption" +
-      "s\022\017\n\007extends\030\001 \003(\t\"\037\n\014OneofOptions\022\017\n\007ex" +
-      "tends\030\001 \003(\t:G\n\007options\022\034.google.protobuf" +
-      ".FileOptions\030\374\007 \001(\0132\027.scalapb.ScalaPbOpt" +
-      "ions:J\n\007message\022\037.google.protobuf.Messag" +
-      "eOptions\030\374\007 \001(\0132\027.scalapb.MessageOptions" +
-      ":D\n\005field\022\035.google.protobuf.FieldOptions" +
-      "\030\374\007 \001(\0132\025.scalapb.FieldOptions:I\n\014enum_o" +
-      "ptions\022\034.google.protobuf.EnumOptions\030\374\007 " +
-      "\001(\0132\024.scalapb.EnumOptions:Q\n\nenum_value\022" +
-      "!.google.protobuf.EnumValueOptions\030\374\007 \001(" +
-      "\0132\031.scalapb.EnumValueOptions:D\n\005oneof\022\035." +
-      "google.protobuf.OneofOptions\030\374\007 \001(\0132\025.sc" +
-      "alapb.OneofOptionsB9\n\030scalapb.options.co" +
-      "mpiler\342?\034\n\030scalapb.options.compiler\020\001"
+      "nsScope\022\024\n\006lenses\030\014 \001(\010:\004true\022\037\n\027retain_" +
+      "source_code_info\030\r \001(\010\022\'\n\035test_only_no_j" +
+      "ava_conversions\030\241\215\006 \001(\010\"%\n\014OptionsScope\022" +
+      "\010\n\004FILE\020\000\022\013\n\007PACKAGE\020\001\"~\n\016MessageOptions" +
+      "\022\017\n\007extends\030\001 \003(\t\022\031\n\021companion_extends\030\002" +
+      " \003(\t\022\023\n\013annotations\030\003 \003(\t\022\014\n\004type\030\004 \001(\t\022" +
+      "\035\n\025companion_annotations\030\005 \003(\t\"\224\001\n\014Field" +
+      "Options\022\014\n\004type\030\001 \001(\t\022\022\n\nscala_name\030\002 \001(" +
+      "\t\022\027\n\017collection_type\030\003 \001(\t\022\020\n\010key_type\030\004" +
+      " \001(\t\022\022\n\nvalue_type\030\005 \001(\t\022\023\n\013annotations\030" +
+      "\006 \003(\t\022\016\n\006no_box\030\036 \001(\010\"G\n\013EnumOptions\022\017\n\007" +
+      "extends\030\001 \003(\t\022\031\n\021companion_extends\030\002 \003(\t" +
+      "\022\014\n\004type\030\003 \001(\t\"#\n\020EnumValueOptions\022\017\n\007ex" +
+      "tends\030\001 \003(\t\"\037\n\014OneofOptions\022\017\n\007extends\030\001" +
+      " \003(\t:G\n\007options\022\034.google.protobuf.FileOp" +
+      "tions\030\374\007 \001(\0132\027.scalapb.ScalaPbOptions:J\n" +
+      "\007message\022\037.google.protobuf.MessageOption" +
+      "s\030\374\007 \001(\0132\027.scalapb.MessageOptions:D\n\005fie" +
+      "ld\022\035.google.protobuf.FieldOptions\030\374\007 \001(\013" +
+      "2\025.scalapb.FieldOptions:I\n\014enum_options\022" +
+      "\034.google.protobuf.EnumOptions\030\374\007 \001(\0132\024.s" +
+      "calapb.EnumOptions:Q\n\nenum_value\022!.googl" +
+      "e.protobuf.EnumValueOptions\030\374\007 \001(\0132\031.sca" +
+      "lapb.EnumValueOptions:D\n\005oneof\022\035.google." +
+      "protobuf.OneofOptions\030\374\007 \001(\0132\025.scalapb.O" +
+      "neofOptionsB9\n\030scalapb.options.compiler\342" +
+      "?\034\n\030scalapb.options.compiler\020\001"
     };
     com.google.protobuf.Descriptors.FileDescriptor.InternalDescriptorAssigner assigner =
         new com.google.protobuf.Descriptors.FileDescriptor.    InternalDescriptorAssigner() {
@@ -8942,7 +9087,7 @@ public final class Scalapb {
     internal_static_scalapb_ScalaPbOptions_fieldAccessorTable = new
       com.google.protobuf.GeneratedMessageV3.FieldAccessorTable(
         internal_static_scalapb_ScalaPbOptions_descriptor,
-        new java.lang.String[] { "PackageName", "FlatPackage", "Import", "Preamble", "SingleFile", "NoPrimitiveWrappers", "PrimitiveWrappers", "CollectionType", "PreserveUnknownFields", "ObjectName", "Scope", "Lenses", "TestOnlyNoJavaConversions", });
+        new java.lang.String[] { "PackageName", "FlatPackage", "Import", "Preamble", "SingleFile", "NoPrimitiveWrappers", "PrimitiveWrappers", "CollectionType", "PreserveUnknownFields", "ObjectName", "Scope", "Lenses", "RetainSourceCodeInfo", "TestOnlyNoJavaConversions", });
     internal_static_scalapb_MessageOptions_descriptor =
       getDescriptor().getMessageTypes().get(1);
     internal_static_scalapb_MessageOptions_fieldAccessorTable = new